% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{GetConvexHull}
\alias{GetConvexHull}
\title{Get convex hull at a given threshold}
\usage{
GetConvexHull(threshold = 2, df, height, x, y)
}
\arguments{
\item{threshold}{What value to exclude numbers worse than}

\item{df}{The data.frame}

\item{height}{The variable name to use for the height threshold}

\item{x}{The first variable to look at for the hull}

\item{y}{The second variable to look at for the hull}
}
\description{
For a given delta lnL, get the convex hull (blob encircling the points) for two dimensions
}
