% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{MergeExhaustiveForPlotting}
\alias{MergeExhaustiveForPlotting}
\title{Function to merge adaptive sampling sims for plotting}
\usage{
MergeExhaustiveForPlotting(exhaustive.object)
}
\arguments{
\item{exhaustive.object}{Return of BMhybExhaustive}
}
\value{
Returns a single BMhyb object with results from all models merged (use for plotting)
}
\description{
Function to merge adaptive sampling sims for plotting
}
