% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nicotiana}
\alias{nicotiana}
\title{Nicotiana dataset}
\format{A list with two items:
\describe{
  \item{phy.graph}{the phylogenetic network in ape::evonet format}
  \item{trait}{a vector of trait data}
}}
\description{
A dataset containing a phylogenetic network and trait data for Nicotiana species
}
\details{
The tree and data come from

Chase M.W., Knapp S., Cox A.V., Clarkson J.J., Butsko Y., Joseph J., Savolainen V., and  Parokonny A.S. 2003. Molecular systematics, GISH and the origin of hybrid taxa in Nicotiana(Solanaceae). Annals of Botany 92: 107-127.

Clarkson J.J., Lim K.Y., Kovarik A., Chase M.W., Knapp S. and Leitch A.R. 2005. Long-term genome diploidization I allopolyploid Nicotiana section Repandae(Solanaceae). New Phytologist 168:241-252.

Komori T., Myers P.N., Yamada S., Kubo T., and Imaseki H. 2000. Comparative study of the Nicotiana species with respect to water deficit tolerance during early growth. Euphytica 116:121-130.
}
\keyword{datasets}
