% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{plot.BMhybExhaustiveResult}
\alias{plot.BMhybExhaustiveResult}
\title{Plot BMhybExhaustive result}
\usage{
\method{plot}{BMhybExhaustiveResult}(x, ...)
}
\arguments{
\item{x}{A BMhybExhaustive object (result of a BMhybExhaustive() call)}

\item{...}{Other parameters to pass to hybResult}
}
\description{
Note this aggregates the info from all the model runs and plots the MLE across all of them and the contours from all the sims
}
