% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{plot.BMhybResult}
\alias{plot.BMhybResult}
\title{Plot BMhyb result}
\usage{
\method{plot}{BMhybResult}(x, style = "univariate", ...)
}
\arguments{
\item{x}{A BMhyb object (result of a BMhyb() call)}

\item{style}{Either univariate or contour}

\item{...}{Other arguments to pass to plot}
}
\description{
Shows the plot of confidence regions with MLEs indicated (red dots) or a plot of pairs of traits together. Note that for the latter plot, it converts the sampled points to an even grid with interpolation; it sets any points with likelihood worse than ten units to just ten units worse so that you can see the colors near the area of the optimum.
}
