% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{plot.evonet}
\alias{plot.evonet}
\title{Plot evonet object}
\usage{
\method{plot}{evonet}(x, ...)
}
\arguments{
\item{x}{An evonet object}

\item{...}{Other arguments to pass to plot}
}
\description{
ape's reorder.evonet function can case C crashes if the edge numbering is not in the manner ape expects (even if obvious requirements like terminal nodes being 1:Ntip and root node being Ntip+1 are met). Instead, we use igraph plotting.
}
