% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{print.BMhybResult}
\alias{print.BMhybResult}
\title{Print BMhyb result}
\usage{
\method{print}{BMhybResult}(x, ...)
}
\arguments{
\item{x}{A BMhyb object (result of a BMhyb() call)}

\item{...}{Other arguments to pass to this function}
}
\description{
Print BMhyb result
}
