% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{summary.BMhybExhaustiveResult}
\alias{summary.BMhybExhaustiveResult}
\title{Summarize BMhybExhaustive result}
\usage{
\method{summary}{BMhybExhaustiveResult}(x)
}
\arguments{
\item{x}{A BMhybExhaustive object (result of a BMhybExhaustive() call)}
}
\description{
Summarize BMhybExhaustive result
}
