% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{summary.BMhybResult}
\alias{summary.BMhybResult}
\title{Summarize BMhyb result}
\usage{
\method{summary}{BMhybResult}(x)
}
\arguments{
\item{x}{A BMhyb object (result of a BMhyb() call)}
}
\description{
Summarize BMhyb result
}
