% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{plot.BMhybResult}
\alias{plot.BMhybResult}
\title{Plot BMhyb result}
\usage{
\method{plot}{BMhybResult}(x, style = "univariate", ...)
}
\arguments{
\item{x}{A BMhyb object (result of a BMhyb() call)}

\item{style}{Either univariate or contour}

\item{...}{Other arguments to pass to plot}
}
\description{
Shows the plot of confidence regions with MLEs indicated (red dots)
}
