% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{plot.BMhybResult}
\alias{plot.BMhybResult}
\title{Plot BMhyb result}
\usage{
plot.BMhybResult(x, ...)
}
\arguments{
\item{x}{A BMhyb object (result of a BMhyb() call)}

\item{...}{Other arguments to pass to plot}
}
\description{
Shows the plot of confidence regions with MLEs indicated (red dots)
}
