% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{print.BMhybExhaustiveResult}
\alias{print.BMhybExhaustiveResult}
\title{Print BMhybExhaustive result}
\usage{
\method{print}{BMhybExhaustiveResult}(x, ...)
}
\arguments{
\item{x}{A BMhybExhaustive object (result of a BMhybExhaustive() call)}

\item{...}{Other arguments to pass to this function}
}
\description{
Print BMhybExhaustive result
}
