% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{summary.BMhybExhaustiveResult}
\alias{summary.BMhybExhaustiveResult}
\title{Summarize BMhybExhaustive result}
\usage{
\method{summary}{BMhybExhaustiveResult}(object, ...)
}
\arguments{
\item{object}{A BMhybExhaustive object (result of a BMhybExhaustive() call)}

\item{...}{Other arguments to pass to this function}
}
\value{
A data.frame with summarized results
}
\description{
Summarize BMhybExhaustive result
}
