% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhyb.r
\name{summary.BMhybResult}
\alias{summary.BMhybResult}
\title{Summarize BMhyb result}
\usage{
\method{summary}{BMhybResult}(object, ...)
}
\arguments{
\item{object}{A BMhyb object (result of a BMhyb() call)}

\item{...}{Other arguments to pass to this function}
}
\value{
A data.frame with summarized results
}
\description{
Summarize BMhyb result
}
